requirements.txt
requests
pandas
unidecode
psycopg2-binary
python-dotenv
